feat: 가계부 캘린더 기능 구현 및 계정/로그인 UI 정리- 가계부 월별 캘린더 조회 기능 구현- 날짜별 수입/지출 데이터 표시 로직 추가- 계정 페이지(Me) UI 구조 개선 (정보/보안/위험 영역 분리)- 로그인 페이지 레이아웃 및 가독성 개선- 전역 CSS(index.css) 정리 및 중복 import/충돌 문제 해결

+++ backend/src/main/java/com/example/cashbook/dto/CustomUserDetails.java
@@ -39,7 +39,8 @@ public class CustomUserDetails implements UserDetails { // SpringSecurity UserDe
         return user.getRole();
     }
 
-    public Long getId() {
-        return user.getId();
-    }
+	public int getId() {
+		return user.getId();
+	}
+
 }
