+++ backend/src/main/java/com/example/cashbook/dto/UserMeResponse.java
@@ -8,7 +8,7 @@ import lombok.Data;
 @Data
 @AllArgsConstructor
 public class UserMeResponse {
-	private Long id; // 사용자 PK
+	private int id; // 사용자 PK
 	private String username; // 아이디
 	private String role; // USER / ADMIN
 	private LocalDateTime createdAt; // 가입일
